Fixed AWS pdf.

+++ generate_pdf_aws.php
@@ -55,32 +55,44 @@ function awsToTex( $aws )
         $abstract = 'Not disclosed yet!';
 
     // Add user image.
-    $imagefile = getLoginPicturePath( $aws['speaker'] );
-
+    $imagefile = getLoginPicturePath( $aws['speaker'], 'null' );
     $imagefile = getThumbnail( $imagefile );
-    echo "Low resolution $imagefile";
 
     $speakerImg = '\includegraphics[width=5cm]{' . $imagefile . '}';
 
     // Date and plate
     $dateAndPlace =  humanReadableDate( $aws[ 'date' ] ) .  
-            ', 4:00pm \faHome \, \textbf{Haapus (LH1), ELC, NCBS}';
+            ', 4:00 pm \faHome \, \textbf{Haapus (LH1), ELC, NCBS}';
     $dateAndPlace = '\faClockO \, ' . $dateAndPlace;
 
-    $head = '\begin{tikzpicture}[ every node/.style={rectangle
-        ,inner sep=1pt,node distance=5mm,text width=0.65\textwidth} ]';
-    $head .= '\node[text width=5cm] (image) at (0,0) {' . $speakerImg . '};';
-    $head .= '\node[above right=of image] (schedule)  {\hfill ' . 
-                $dateAndPlace . '};';
-    $head .= '\node[right=of image] (title) { ' .  '\textsc{\LARGE ' . $title . '} };';
-    $head .= '\node[below=of title] (author) { ' .  '{' . $speaker . '} };';
+    // Two columns here.
+    $head = '';
+    $head .= '\begin{minipage}[b][6cm][b]{\linewidth} ';
+
+    // Header 
+    $head .= '\begin{tikzpicture}[remember picture,overlay,every node/.style={rectangle, node distance=5mm,inner sep=0mm} ]';
+    $head .= '\node[] (ncbs) at ([xshift=-40mm,yshift=-15mm]current page.north east) 
+        { \includegraphics[height=1.5cm]{./data/ncbs_logo.png} };';
+    $head .= '\node[] (instem) at ([xshift=30mm,yshift=-15mm]current page.north west) 
+        { \includegraphics[height=1.5cm]{./data/inStem_logo.png} };';
+    $head .= '\node[ ] (aws) at ($(ncbs)!0.5!(instem)$) {\color{blue}Annual Work Seminar};';
+    $head .= '\draw[dotted,thick] ([yshift=-5mm]ncbs.south) -- ([yshift=-5mm]instem.south)
+                node[above,midway] {\color{blue} ' . $dateAndPlace . ' };';
     $head .= '\end{tikzpicture}';
 
-
+    $head .= '\par';
+    $head .= '\begin{tikzpicture}[every node/.style={rectangle, node distance=5mm,inner sep=0mm} ]';
+    $head .= '\node (align) at (0,0) {};';
+    $head .= '\node[left=of align] (img) { ' . $speakerImg . '};';
+    $head .= '\node[above right=of img] (date) { };';
+    $head .= '\node[right=of img,text width=0.65\linewidth] (title) {\textsc{\Large ' . $title . '}};';
+    $head .= '\node[below=of title,text width=0.65\linewidth] (author) {\textbf{' . $speaker . '}};';
+    $head .= '\end{tikzpicture}';
+    $head .= '\end{minipage}';
 
     // Header
     $tex = array( $head );
-    $tex[] = '\par';
+    $tex[] = '\par \vspace{3mm}';
 
     // remove html formating before converting to tex.
     $tempFile = tempnam( "/tmp", "hippo_abstract" );
@@ -100,9 +112,10 @@ function awsToTex( $aws )
     $extra .= '\textbf{Thesis Committee Member(s)} & ' . implode( ", ", $tcm ) . '\\\\';
     $extra .= '\end{tabular}';
     $extra .= '\end{table}';
-
     $tex[] = $extra;
+
     return implode( "\n", $tex );
+
 } // Function ends.
 
 if( ! array_key_exists( 'date', $_GET ) )
@@ -125,23 +138,24 @@ else
 
 // Intialize pdf template.
 $tex = array( "\documentclass[]{article}"
-    , "\usepackage[margin=20mm,top=3cm,a4paper]{geometry}"
+    , "\usepackage[margin=25mm,top=20mm,a4paper]{geometry}"
     , "\usepackage[]{graphicx}"
     , "\usepackage[]{grffile}"
     , "\usepackage[]{amsmath,amssymb}"
     , "\usepackage[]{color}"
     , "\usepackage{tikz}"
     , "\usepackage{fontawesome}"
-    , '\usepackage{fancyhdr}'
-    , '\linespread{1.2}'
-    , '\pagestyle{fancy}'
     , '\pagenumbering{gobble}'
-    , '\lhead{\textsc{{\color{blue} Annual Work Seminar}}}'
-    //, '\rhead{National Center for Biological Sciences, Bangalore \\linebreak
-        //TATA Institute of Fundamental Research, Mumbai}'
+    //, '\usepackage{fancyhdr}'
+    , '\linespread{1.2}'
+    //, '\pagestyle{fancy}'
+    //, '\chead{\textsc{{\color{blue} Annual Work Seminar}}}'
     , '\usetikzlibrary{calc,positioning,arrows}'
-    , '\usepackage[]{ebgaramond}'
     , '\usepackage[T1]{fontenc}'
+    //, '\usepackage[]{ebgaramond}'
+    , '\usepackage[sfdefault,light]{FiraSans}'
+    //, '\rhead{ \includegraphics[height=15 mm]{./data/ncbs_logo.png} }'
+    //, '\lhead { \includegraphics[height=15 mm]{./data/inStem_logo.png}}'
     , '\begin{document}'
     );
 
@@ -152,9 +166,8 @@ foreach( $awses as $aws )
     $tex[] = awsToTex( $aws );
 
     // If speaker has instem id, put InSTEM as well in header.
-    $inst = emailInstitute( getLoginEmail( $aws[ 'speaker' ] ), "latex" );
-    $tex[] .= '\rhead{ \textsc { ' . $inst . '}}';
-    $tex[] = '\pagebreak';
+    //$inst = emailInstitute( getLoginEmail( $aws[ 'speaker' ] ), "latex" );
+    //$tex[] = '\newpage';
 }
 
 $tex[] = '\end{document}';
@@ -173,6 +186,8 @@ else
 
 // Remove tex from the end and apped pdf.
 $pdfFile = rtrim( $texFile, 'tex' ) . 'pdf';
+if( file_exists( $pdfFile ) )
+    unlink( $pdfFile );
 
 file_put_contents( $texFile,  $TeX );
 if( file_exists( $texFile ) )
@@ -197,7 +212,10 @@ else
     echo "<pre> $res </pre>";
 }
 
-unlink( $texFile );
+if( file_exists( $texFile ) )
+{
+    unlink( $texFile );
+}
 
 echo "<br/>";
 echo closePage( );
